perf: Job 切换 GSE2.0 支持灰度策略 #2461
