+++ src/backend/commons/common-service/src/main/java/com/tencent/bk/job/common/service/ConfigRefreshEventListener.java
@@ -24,9 +24,10 @@
 
 package com.tencent.bk.job.common.service;
 
-import com.tencent.bk.job.common.util.feature.FeatureToggle;
+import com.tencent.bk.job.common.util.feature.FeatureStore;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.context.environment.EnvironmentChangeEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
@@ -40,7 +41,12 @@ import java.util.StringJoiner;
 @Component
 @Slf4j
 public class ConfigRefreshEventListener {
-    public ConfigRefreshEventListener() {
+
+    private final FeatureStore featureStore;
+
+    @Autowired
+    public ConfigRefreshEventListener(FeatureStore featureStore) {
+        this.featureStore = featureStore;
         log.info("Init ConfigRefreshEventListener");
     }
 
@@ -76,7 +82,7 @@ public class ConfigRefreshEventListener {
         boolean isFeatureToggleConfigChanged =
             changedKeys.stream().anyMatch(changedKey -> changedKey.startsWith("job.features."));
         if (isFeatureToggleConfigChanged) {
-            FeatureToggle.reload();
+            featureStore.load(true);
         }
     }
 }
