+++ src/backend/commons/common-service/src/main/java/com/tencent/bk/job/common/service/feature/FeatureIdConstants.java
@@ -22,23 +22,27 @@
  * IN THE SOFTWARE.
  */
 
-package com.tencent.bk.job.logsvr;
+package com.tencent.bk.job.common.service.feature;
 
-import com.tencent.bk.job.common.service.boot.JobBootApplication;
-import com.tencent.bk.job.common.service.feature.FeatureToggleConfig;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.availability.ApplicationAvailabilityAutoConfiguration;
-import org.springframework.boot.autoconfigure.jooq.JooqAutoConfiguration;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.cloud.openfeign.EnableFeignClients;
+/**
+ * 特性ID定义
+ */
+public interface FeatureIdConstants {
+    /**
+     * 特性: 对接 GSE2.0
+     */
+    String FEATURE_GSE_V2 = "gseV2";
+    /**
+     * 特性: OpenAPI 兼容bk_biz_id参数
+     */
+    String FEATURE_BK_BIZ_ID_COMPATIBLE = "bkBizIdCompatible";
+    /**
+     * 特性-第三方文件源
+     */
+    String FEATURE_FILE_MANAGE = "fileManage";
 
-@JobBootApplication(
-    scanBasePackages = "com.tencent.bk.job.logsvr",
-    exclude = {JooqAutoConfiguration.class, ApplicationAvailabilityAutoConfiguration.class})
-@EnableFeignClients
-@EnableConfigurationProperties({FeatureToggleConfig.class})
-public class JobLogBootApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(JobLogBootApplication.class, args);
-    }
+    /**
+     * 特性-是否支持GSE 获取文件分发任务结果的API协议(2.0版本之前)
+     */
+    String GSE_FILE_PROTOCOL_BEFORE_V2 = "gseFileProtocolBeforeV2";
 }
