+++ src/backend/commons/common/src/main/java/com/tencent/bk/job/common/util/feature/ToggleStrategy.java
@@ -22,41 +22,46 @@
  * IN THE SOFTWARE.
  */
 
-package com.tencent.bk.job.common.util.feature.strategy;
+package com.tencent.bk.job.common.util.feature;
 
-import com.tencent.bk.job.common.util.feature.FeatureExecutionContext;
-
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-public class AnyMatchToggleStrategy extends AbstractToggleStrategy {
+/**
+ * 特性开关开启策略
+ */
+public interface ToggleStrategy {
+    /**
+     * 获取特性开关开启策略ID
+     */
+    String getId();
 
     /**
-     * 特性开关开启策略ID
+     * 获取策略说明
      */
-    public static final String STRATEGY_ID = "AnyMatchToggleStrategy";
-
-    public AnyMatchToggleStrategy(String description,
-                                  List<ToggleStrategy> strategies,
-                                  Map<String, String> initParams) {
-        super(STRATEGY_ID, description, strategies, initParams);
-        assertRequiredAtLeastOneStrategy();
+    default String getDescription() {
+        return "";
     }
 
-    @Override
-    public boolean evaluate(String featureId, FeatureExecutionContext ctx) {
-        assertRequiredAtLeastOneStrategy();
-        for (ToggleStrategy strategy : getCompositeToggleStrategies()) {
-            boolean isMatch = strategy.evaluate(featureId, ctx);
-            if (isMatch) {
-                return true;
-            }
-        }
-        return false;
+    /**
+     * 获取初始化参数
+     */
+    Map<String, String> getInitParams();
+
+    /**
+     * 获取复合策略
+     */
+    default List<ToggleStrategy> getCompositeToggleStrategies() {
+        return Collections.emptyList();
     }
 
-    @Override
-    public List<ToggleStrategy> getCompositeToggleStrategies() {
-        return this.compositeStrategies;
-    }
+    /**
+     * 判断是否开启特性
+     *
+     * @param featureId 特性ID
+     * @param ctx       执行上下文
+     * @return true: 特性开启,false: 特性关闭
+     */
+    boolean evaluate(String featureId, FeatureExecutionContext ctx);
 }
